Class MariaDBDriver with constructor MariaDB.

An MariaDB driver implementing the R database (DBI) API. This class should always be initialized with the MariaDB() function. It returns a singleton that allows you to connect to MariaDB.

Usage

# S4 method for class 'MariaDBDriver'
dbGetInfo(dbObj, ...)

# S4 method for class 'MariaDBDriver'
dbIsValid(dbObj, ...)

# S4 method for class 'MariaDBDriver'
dbUnloadDriver(drv, ...)